Moved animations to dedicated files
