+++ example/src/animations.ts
@@ -1,0 +1,5 @@
+export default {
+    type: "slide",
+    direction: "right",
+    duration: 350,
+} as const;
