+++ example/src/common/constants.ts
@@ -1,0 +1,9 @@
+import { AnimationConfig, AnimationKeyframeEffectConfig } from "@react-motion-router/core";
+import { iOS, isPWA } from "./utils";
+
+export const STATIC_ANIMATION: AnimationKeyframeEffectConfig | AnimationConfig = {
+  keyframes: [],
+  options: {
+    duration: iOS() && !isPWA() ? 0 : 350
+  }
+};
